@@ -1,32 +1,104 @@
 package filelocker
 
 import (
-	"io/ioutil"
+	"context"
+	"os"
 	"testing"
+	"time"
 
 	"github.com/stretchr/testify/assert"
-	"github.com/tus/tusd/pkg/handler"
+	"github.com/tus/tusd/v2/pkg/handler"
 )
 
 var _ handler.Locker = &FileLocker{}
 
-func TestFileLocker(t *testing.T) {
+func TestMemoryLocker_LockAndUnlock(t *testing.T) {
 	a := assert.New(t)
 
-	dir, err := ioutil.TempDir("", "tusd-file-locker")
+	dir, err := os.MkdirTemp("", "tusd-file-locker")
 	a.NoError(err)
 
-	locker := FileLocker{dir}
+	locker := New(dir)
 
 	lock1, err := locker.NewLock("one")
 	a.NoError(err)
 
-	a.NoError(lock1.Lock())
-	a.Equal(handler.ErrFileLocked, lock1.Lock())
+	a.NoError(lock1.Lock(context.Background(), func() {
+		panic("must not be called")
+	}))
+	a.NoError(lock1.Unlock())
+
+	// Ensure that directory is empty
+	assertEmptyDirectory(dir, a)
+}
+
+func TestFileLocker_Timeout(t *testing.T) {
+	a := assert.New(t)
+
+	dir, err := os.MkdirTemp("", "tusd-file-locker")
+	a.NoError(err)
+
+	locker := New(dir)
+
+	lock1, err := locker.NewLock("one")
+	a.NoError(err)
+	a.NoError(lock1.Lock(context.Background(), func() {}))
+
+	ctx, cancel := context.WithTimeout(context.Background(), time.Millisecond)
+	defer cancel()
 
 	lock2, err := locker.NewLock("one")
 	a.NoError(err)
-	a.Equal(handler.ErrFileLocked, lock2.Lock())
+	err = lock2.Lock(ctx, func() {
+		panic("must not be called")
+	})
+	a.Equal(err, handler.ErrLockTimeout)
 
 	a.NoError(lock1.Unlock())
+
+	// Ensure that directory is empty
+	assertEmptyDirectory(dir, a)
+}
+
+func TestMemoryLocker_RequestUnlock(t *testing.T) {
+	a := assert.New(t)
+
+	dir, err := os.MkdirTemp("", "tusd-file-locker")
+	a.NoError(err)
+
+	locker := New(dir)
+	locker.AcquirerPollInterval = 100 * time.Millisecond
+	locker.HolderPollInterval = 300 * time.Millisecond
+	releaseRequestCalled := false
+
+	lock1, err := locker.NewLock("one")
+	a.NoError(err)
+	a.NoError(lock1.Lock(context.Background(), func() {
+		releaseRequestCalled = true
+		<-time.After(10 * time.Millisecond)
+		a.NoError(lock1.Unlock())
+	}))
+
+	ctx, cancel := context.WithTimeout(context.Background(), time.Second)
+	defer cancel()
+
+	lock2, err := locker.NewLock("one")
+	a.NoError(err)
+	a.NoError(lock2.Lock(ctx, func() {
+		panic("must not be called")
+	}))
+	a.NoError(lock2.Unlock())
+
+	a.True(releaseRequestCalled)
+
+	// Ensure that directory is empty
+	assertEmptyDirectory(dir, a)
+}
+
+func assertEmptyDirectory(dir string, a *assert.Assertions) {
+	file, err := os.Open(dir)
+	a.NoError(err)
+	entries, err := file.Readdirnames(0)
+	a.NoError(err)
+	a.Equal(0, len(entries))
 }
