@@ -3,13 +3,13 @@ package filestore
 import (
 	"context"
 	"io"
-	"io/ioutil"
+	"os"
 	"path/filepath"
 	"strings"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
-	"github.com/tus/tusd/pkg/handler"
+	"github.com/tus/tusd/v2/pkg/handler"
 )
 
 // Test interface implementation of Filestore
@@ -21,7 +21,7 @@ var _ handler.LengthDeferrerDataStore = FileStore{}
 func TestFilestore(t *testing.T) {
 	a := assert.New(t)
 
-	tmp, err := ioutil.TempDir("", "tusd-filestore-")
+	tmp, err := os.MkdirTemp("", "tusd-filestore-")
 	a.NoError(err)
 
 	store := FileStore{tmp}
@@ -62,7 +62,7 @@ func TestFilestore(t *testing.T) {
 	reader, err := upload.GetReader(ctx)
 	a.NoError(err)
 
-	content, err := ioutil.ReadAll(reader)
+	content, err := io.ReadAll(reader)
 	a.NoError(err)
 	a.Equal("hello world", string(content))
 	reader.(io.Closer).Close()
@@ -103,7 +103,7 @@ func TestNotFound(t *testing.T) {
 func TestConcatUploads(t *testing.T) {
 	a := assert.New(t)
 
-	tmp, err := ioutil.TempDir("", "tusd-filestore-concat-")
+	tmp, err := os.MkdirTemp("", "tusd-filestore-concat-")
 	a.NoError(err)
 
 	store := FileStore{tmp}
@@ -152,7 +152,7 @@ func TestConcatUploads(t *testing.T) {
 	reader, err := finUpload.GetReader(ctx)
 	a.NoError(err)
 
-	content, err := ioutil.ReadAll(reader)
+	content, err := io.ReadAll(reader)
 	a.NoError(err)
 	a.Equal("abcdefghi", string(content))
 	reader.(io.Closer).Close()
@@ -161,7 +161,7 @@ func TestConcatUploads(t *testing.T) {
 func TestDeclareLength(t *testing.T) {
 	a := assert.New(t)
 
-	tmp, err := ioutil.TempDir("", "tusd-filestore-declare-length-")
+	tmp, err := os.MkdirTemp("", "tusd-filestore-declare-length-")
 	a.NoError(err)
 
 	store := FileStore{tmp}
