@@ -3,8 +3,9 @@ package cli
 import (
 	"net/http"
 
-	"github.com/tus/tusd/pkg/handler"
-	"github.com/tus/tusd/pkg/prometheuscollector"
+	"github.com/tus/tusd/v2/pkg/handler"
+	"github.com/tus/tusd/v2/pkg/hooks"
+	"github.com/tus/tusd/v2/pkg/prometheuscollector"
 
 	"github.com/prometheus/client_golang/prometheus"
 	"github.com/prometheus/client_golang/prometheus/promhttp"
@@ -15,19 +16,12 @@ var MetricsOpenConnections = prometheus.NewGauge(prometheus.GaugeOpts{
 	Help: "Current number of open connections.",
 })
 
-var MetricsHookErrorsTotal = prometheus.NewCounterVec(
-	prometheus.CounterOpts{
-		Name: "tusd_hook_errors_total",
-		Help: "Total number of execution errors per hook type.",
-	},
-	[]string{"hooktype"},
-)
-
-func SetupMetrics(handler *handler.Handler) {
+func SetupMetrics(mux *http.ServeMux, handler *handler.Handler) {
 	prometheus.MustRegister(MetricsOpenConnections)
-	prometheus.MustRegister(MetricsHookErrorsTotal)
+	prometheus.MustRegister(hooks.MetricsHookErrorsTotal)
+	prometheus.MustRegister(hooks.MetricsHookInvocationsTotal)
 	prometheus.MustRegister(prometheuscollector.New(handler.Metrics))
 
 	stdout.Printf("Using %s as the metrics path.\n", Flags.MetricsPath)
-	http.Handle(Flags.MetricsPath, promhttp.Handler())
+	mux.Handle(Flags.MetricsPath, promhttp.Handler())
 }
