@@ -14,13 +14,15 @@ import (
 // In addition, the bodyReader keeps track of how many bytes were read.
 type bodyReader struct {
 	reader       io.Reader
+	closer       io.Closer
 	err          error
 	bytesCounter int64
 }
 
-func newBodyReader(r io.Reader) *bodyReader {
+func newBodyReader(r io.ReadCloser, maxSize int64) *bodyReader {
 	return &bodyReader{
-		reader: r,
+		reader: io.LimitReader(r, maxSize),
+		closer: r,
 	}
 }
 
@@ -31,13 +33,24 @@ func (r *bodyReader) Read(b []byte) (int, error) {
 
 	n, err := r.reader.Read(b)
 	atomic.AddInt64(&r.bytesCounter, int64(n))
-	r.err = err
+	if err != nil {
+		// We can ignore some of these errors:
+		// - io.EOF means that the request body was fully read
+		// - io.ErrClosedPipe means that the bodyReader closed the request body because the upload is
+		//   is stopped or the server shuts down.
+		// - io.UnexpectedEOF means that the client aborted the request.
+		// In all of those cases, we do not forward the error to the storage,
+		// but act like the body just ended naturally.
+		if err == io.EOF || err == io.ErrClosedPipe || err == io.ErrUnexpectedEOF {
+			return n, io.EOF
+		}
 
-	if err == io.EOF {
-		return n, io.EOF
-	} else {
-		return n, nil
+		// Other errors are stored for retrival with hasError, but is not returned
+		// to the consumer.
+		r.err = err
 	}
+
+	return n, nil
 }
 
 func (r bodyReader) hasError() error {
@@ -50,4 +63,9 @@ func (r bodyReader) hasError() error {
 
 func (r *bodyReader) bytesRead() int64 {
 	return atomic.LoadInt64(&r.bytesCounter)
+}
+
+func (r *bodyReader) closeWithError(err error) {
+	r.closer.Close()
+	r.err = err
 }
