@@ -7,7 +7,9 @@ import (
 
 type MetaData map[string]string
 
+// FileInfo contains information about a single upload resource.
 type FileInfo struct {
+	// ID is the unique identifier of the upload resource.
 	ID string
 	// Total file size in bytes specified in the NewUpload call
 	Size int64
@@ -31,53 +33,79 @@ type FileInfo struct {
 	// store is used. This map may also be nil.
 	Storage map[string]string
 
-	// stopUpload is the cancel function for the upload's context.Context. When
-	// invoked it will interrupt the writes to DataStore#WriteChunk.
-	stopUpload context.CancelFunc
+	// stopUpload is a channel for communicating that an upload should by stopped
+	// and interrupt the writes to DataStore#WriteChunk.
+	stopUpload chan HTTPResponse
 }
 
-// StopUpload interrupts an running upload from the server-side. This means that
+// StopUpload interrupts a running upload from the server-side. This means that
 // the current request body is closed, so that the data store does not get any
 // more data. Furthermore, a response is sent to notify the client of the
 // interrupting and the upload is terminated (if supported by the data store),
-// so the upload cannot be resumed anymore.
-func (f FileInfo) StopUpload() {
+// so the upload cannot be resumed anymore. The response to the client can be
+// optionally modified by providing values in the HTTPResponse struct.
+func (f FileInfo) StopUpload(response HTTPResponse) {
 	if f.stopUpload != nil {
-		f.stopUpload()
+		f.stopUpload <- response
 	}
+}
+
+// FileInfoChanges collects changes the should be made to a FileInfo struct. This
+// can be done using the PreUploadCreateCallback to modify certain properties before
+// an upload is created. Properties which should not be modified (e.g. Size or Offset)
+// are intentionally left out here.
+type FileInfoChanges struct {
+	// If ID is not empty, it will be passed to the data store, allowing
+	// hooks to influence the upload ID. Be aware that a data store is not required to
+	// respect a pre-defined upload ID and might overwrite or modify it. However,
+	// all data stores in the github.com/tus/tusd package do respect pre-defined IDs.
+	ID string
+
+	// If MetaData is not nil, it replaces the entire user-defined meta data from
+	// the upload creation request. You can add custom meta data fields this way
+	// or ensure that only certain fields from the user-defined meta data are saved.
+	// If you want to retain only specific entries from the user-defined meta data, you must
+	// manually copy them into this MetaData field.
+	// If you do not want to store any meta data, set this field to an empty map (`MetaData{}`).
+	// If you want to keep the entire user-defined meta data, set this field to nil.
+	MetaData MetaData
+
+	// If Storage is not nil, it is passed to the data store to allow for minor adjustments
+	// to the upload storage (e.g. destination file name). The details are specific for each
+	// data store and should be looked up in their respective documentation.
+	// Please be aware that this behavior is currently not supported by any data store in
+	// the github.com/tus/tusd package.
+	Storage map[string]string
 }
 
 type Upload interface {
 	// Write the chunk read from src into the file specified by the id at the
 	// given offset. The handler will take care of validating the offset and
-	// limiting the size of the src to not overflow the file's size. It may
-	// return an os.ErrNotExist which will be interpreted as a 404 Not Found.
-	// It will also lock resources while they are written to ensure only one
+	// limiting the size of the src to not overflow the file's size.
+	// The handler will also lock resources while they are written to ensure only one
 	// write happens per time.
 	// The function call must return the number of bytes written.
 	WriteChunk(ctx context.Context, offset int64, src io.Reader) (int64, error)
 	// Read the fileinformation used to validate the offset and respond to HEAD
-	// requests. It may return an os.ErrNotExist which will be interpreted as a
-	// 404 Not Found.
+	// requests.
 	GetInfo(ctx context.Context) (FileInfo, error)
-	// GetReader returns a reader which allows iterating of the content of an
-	// upload specified by its ID. It should attempt to provide a reader even if
-	// the upload has not been finished yet but it's not required.
-	// If the returned reader also implements the io.Closer interface, the
-	// Close() method will be invoked once everything has been read.
-	// If the given upload could not be found, the error tusd.ErrNotFound should
-	// be returned.
-	GetReader(ctx context.Context) (io.Reader, error)
+	// GetReader returns an io.ReadCloser which allows iterating of the content of an
+	// upload. It should attempt to provide a reader even if the upload has not
+	// been finished yet but it's not required.
+	GetReader(ctx context.Context) (io.ReadCloser, error)
 	// FinisherDataStore is the interface which can be implemented by DataStores
 	// which need to do additional operations once an entire upload has been
 	// completed. These tasks may include but are not limited to freeing unused
 	// resources or notifying other services. For example, S3Store uses this
 	// interface for removing a temporary object.
-	// FinishUpload executes additional operations for the finished upload which
-	// is specified by its ID.
 	FinishUpload(ctx context.Context) error
 }
 
+// DataStore is the base interface for storages to implement. It provides functions
+// to create new uploads and fetch existing ones.
+//
+// Note: the context values passed to all functions is not the request's context,
+// but a similar context. See HookEvent.Context for more details.
 type DataStore interface {
 	// Create a new upload using the size as the file's length. The method must
 	// return an unique id which is used to identify the upload. If no backend
@@ -85,12 +113,14 @@ type DataStore interface {
 	// generate one. The properties Size and MetaData will be filled.
 	NewUpload(ctx context.Context, info FileInfo) (upload Upload, err error)
 
+	// GetUpload fetches the upload with a given ID. If no such upload can be found,
+	// ErrNotFound must be returned.
 	GetUpload(ctx context.Context, id string) (upload Upload, err error)
 }
 
 type TerminatableUpload interface {
-	// Terminate an upload so any further requests to the resource, both reading
-	// and writing, must return os.ErrNotExist or similar.
+	// Terminate an upload so any further requests to the upload resource will
+	// return the ErrNotFound error.
 	Terminate(ctx context.Context) error
 }
 
@@ -146,11 +176,15 @@ type Locker interface {
 type Lock interface {
 	// Lock attempts to obtain an exclusive lock for the upload specified
 	// by its id.
-	// If this operation fails because the resource is already locked, the
-	// tusd.ErrFileLocked must be returned. If no error is returned, the attempt
-	// is consider to be successful and the upload to be locked until UnlockUpload
-	// is invoked for the same upload.
-	Lock() error
+	// If the lock can be acquired, it will return without error. The requestUnlock
+	// callback is invoked when another caller attempts to create a lock. In this
+	// case, the holder of the lock should attempt to release the lock as soon
+	// as possible
+	// If the lock is already held, the holder's requestUnlock function will be
+	// invoked to request the lock to be released. If the context is cancelled before
+	// the lock can be acquired, ErrLockTimeout will be returned without acquiring
+	// the lock.
+	Lock(ctx context.Context, requestUnlock func()) error
 	// Unlock releases an existing lock for the given upload.
 	Unlock() error
 }
