@@ -12,7 +12,7 @@ import (
 	"strconv"
 	"sync/atomic"
 
-	"github.com/tus/tusd/pkg/handler"
+	"github.com/tus/tusd/v2/pkg/handler"
 
 	"github.com/prometheus/client_golang/prometheus"
 )
@@ -25,7 +25,7 @@ var (
 	errorsTotalDesc = prometheus.NewDesc(
 		"tusd_errors_total",
 		"Total number of errors per status.",
-		[]string{"status", "message"}, nil)
+		[]string{"status", "code"}, nil)
 	bytesReceivedDesc = prometheus.NewDesc(
 		"tusd_bytes_received",
 		"Number of bytes received for uploads.",
@@ -55,7 +55,7 @@ func New(metrics handler.Metrics) Collector {
 	}
 }
 
-func (_ Collector) Describe(descs chan<- *prometheus.Desc) {
+func (Collector) Describe(descs chan<- *prometheus.Desc) {
 	descs <- requestsTotalDesc
 	descs <- errorsTotalDesc
 	descs <- bytesReceivedDesc
@@ -79,8 +79,8 @@ func (c Collector) Collect(metrics chan<- prometheus.Metric) {
 			errorsTotalDesc,
 			prometheus.CounterValue,
 			float64(atomic.LoadUint64(valuePtr)),
-			strconv.Itoa(httpError.StatusCode()),
-			httpError.Error(),
+			strconv.Itoa(httpError.StatusCode),
+			httpError.ErrorCode,
 		)
 	}
 
